Fix overflow resulting in infinite loop.

+++ src/driver/cpu/x64/std/memset.cpp
@@ -2,7 +2,7 @@
 
 void memset(void * inPtr, uint8_t inValue, size_t inNum) {
     // Simple, inefficient, but can be upgraded later.
-    for(uint8_t i = 0; i < inNum; i++) {
+    for(size_t i = 0; i < inNum; i++) {
         ((uint8_t *)inPtr)[i] = inValue;
     }
 }
